Move partitioning executors to the Storage module
Move the partitioning executors to new files in the Storage module.
