+++ pyanaconda/modules/storage/partitioning/automatic.py
@@ -27,7 +27,8 @@ from pyanaconda.modules.storage.partitioning.base import PartitioningModule
 from pyanaconda.modules.storage.partitioning.automatic_interface import AutoPartitioningInterface
 from pyanaconda.modules.storage.partitioning.configure import StorageConfigureTask
 from pyanaconda.modules.storage.partitioning.validate import StorageValidateTask
-from pyanaconda.storage.execution import AutomaticPartitioningExecutor
+from pyanaconda.modules.storage.partitioning.automatic_execution import \
+    AutomaticPartitioningExecutor
 
 log = get_module_logger(__name__)
