+++ pyanaconda/modules/storage/partitioning/blivet.py
@@ -25,7 +25,7 @@ from pyanaconda.modules.storage.partitioning.blivet_interface import \
     BlivetPartitioningInterface
 from pyanaconda.modules.storage.partitioning.configure import StorageConfigureTask
 from pyanaconda.modules.storage.partitioning.validate import StorageValidateTask
-from pyanaconda.storage.execution import InteractivePartitioningExecutor
+from pyanaconda.modules.storage.partitioning.base_execution import InteractivePartitioningExecutor
 
 log = get_module_logger(__name__)
