+++ tests/nosetests/pyanaconda_tests/module_blivet_test.py
@@ -30,7 +30,7 @@ from pyanaconda.modules.storage.partitioning.blivet_interface import \
     BlivetPartitioningInterface
 from pyanaconda.modules.storage.partitioning.configure import StorageConfigureTask
 from pyanaconda.modules.storage.partitioning.validate import StorageValidateTask
-from pyanaconda.storage.execution import InteractivePartitioningExecutor
+from pyanaconda.modules.storage.partitioning.base_execution import InteractivePartitioningExecutor
 
 
 class BlivetPartitioningInterfaceTestCase(unittest.TestCase):
